feat: add LoginForm, ForgotPassword with validations and checkbox selection logic
